Repair database and OneToMany and ManyToMany connections in User and Role entities

+++ src/main/java/lits/com/springboot/repository/CityRepository.java
@@ -3,8 +3,8 @@ package lits.com.springboot.repository;
 import lits.com.springboot.model.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepository extends JpaRepository<City, Integer> {
-    City findById(Integer id);
+public interface CityRepository extends JpaRepository<City, Long> {
+    City findById(Long id);
     City findByName(String name);
 
 }
