+++ src/main/java/lits/com/springboot/repository/PersonRepository.java
@@ -5,13 +5,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PersonRepository extends JpaRepository<Person, Integer> {
+public interface PersonRepository extends JpaRepository<Person, Long> {
 
    List<Person> findByNameAndAge(String name, Integer age);
 
    List<Person> findAllByNameContains(String name);
 
-   List<Person> findByCityId(Integer cityId);
+   List<Person> findByCityId(Long cityId);
 
 //   List<Person> findAllByAgeBetweenAndNameContaining(Integer ageFrom, Integer ageTo, String name);
 }
