+++ src/main/java/lits/com/springboot/service/impl/DeadPersonService.java
@@ -21,7 +21,7 @@ public class DeadPersonService implements PersonService {
     private ModelMapper modelMapper;
 
     @Override
-    public PersonDto getById(Integer id) {
+    public PersonDto getById(Long id) {
         Person person = personRepository.findOne(id);
         person = person.getDead() ? person : null;
         return modelMapper.map(person, PersonDto.class);
@@ -29,7 +29,7 @@ public class DeadPersonService implements PersonService {
 
 
     @Override
-    public List<PersonDto> getAllPersonsByCity(Integer cityId) {
+    public List<PersonDto> getAllPersonsByCity(Long cityId) {
         List<PersonDto> personDtos = new ArrayList<>();
         List<Person> persons = personRepository.findByCityId(cityId);
         for (Person person : persons) {
